Use even bisection instead of greedy strategy for picking best square.
In principle picking the probability closest to 0.5 for the next move
should result in fewer moves since it enables you to exclude more layouts.
I'm not sure how common this case is (for that the probability of a square would
need to be more than 0.5), but it should in principle perform better.