Make third kill management more consistent
The third kill would previously not be input when the ninth hit was set
by clicking, which would be confusing for a user who never thought about
the third kill. It would be equally confusing that clearing the ninth
hit without using undo resulted in an invalid state, though that case
was probably never encountered.